Fixes & Updates
- Otimizado mais funcionalidades relacionadas as montarias
- Adicionado item GLASSES que permite ver quais blocos proximos contem
itens
- Fixado bug do hurry up que parava de cair 1 tile antes do previsto, o
que permitia que um ou mais jogadores ficassem vivos nesse tile
- Fixado bug onde o bloco do hurry up podia, em certas circunstâncias,
não matar instantaneamente o player ao cair em cima dele
- Editor de FrameSet reformulado, para trabalhar com o FrameSet em tempo
real editando as tags através de texto

+++ src/frameset_tags/SetEntitySpeed.java
@@ -18,7 +18,7 @@ public class SetEntitySpeed extends FrameTag {
 	public SetEntitySpeed(String tags) {
 		String[] params = validateStringTags(this, tags, 1);
 		try {
-			value = Integer.parseInt(params[0]);
+			value = Double.parseDouble(params[0]);
 		}
 		catch (Exception e) {
 			throw new RuntimeException(params[0] + " - Invalid parameter");
